+++ src/js/services/start-page.js
@@ -3,10 +3,12 @@ const cardEvent = document.querySelector('.gallery');
 const eventsApiService = new EventsApiService();
 import eventsListTpl from '../../templates/card-list.hbs';
 import createsDownloadList from './nev-render-list.js';
+import getPage from './get-page.js';
 
 function onSearch() {
     try {
         fetchRefs();
+
 
     } catch (error) {}
 
@@ -26,4 +28,5 @@ async function fetchRefs() {
     const list = _embedded.events;
     const newFetchEventList = createsDownloadList(list);
     renderEventList(newFetchEventList);
+    getPage(eventsApiService);
 }
